+++ web/apps/photos/src/components/Collections/CollectionInfoWithOptions.tsx
@@ -5,7 +5,7 @@ import LinkIcon from "@mui/icons-material/Link";
 import PeopleIcon from "@mui/icons-material/People";
 import { SetCollectionNamerAttributes } from "components/Collections/CollectionNamer";
 import CollectionOptions from "components/Collections/CollectionOptions";
-import { CollectionSummaryType } from "constants/collection";
+import { CollectionSummaryType } from "utils/collection";
 import type { Dispatch, SetStateAction } from "react";
 import { Collection, CollectionSummary } from "types/collection";
 import { SetFilesDownloadProgressAttributesCreator } from "types/gallery";
