+++ web/apps/photos/src/components/Collections/CollectionSelector/index.tsx
@@ -2,11 +2,6 @@ import { FlexWrapper } from "@ente/shared/components/Container";
 import DialogTitleWithCloseButton from "@ente/shared/components/DialogBox/TitleWithCloseButton";
 import { DialogContent, useMediaQuery } from "@mui/material";
 import { AllCollectionDialog } from "components/Collections/AllCollections/dialog";
-import {
-    COLLECTION_SORT_ORDER,
-    CollectionSummaryType,
-    DUMMY_UNCATEGORIZED_COLLECTION,
-} from "constants/collection";
 import { t } from "i18next";
 import { useEffect, useState } from "react";
 import { createUnCategorizedCollection } from "services/collectionService";
@@ -17,6 +12,9 @@ import {
 } from "types/collection";
 import { CollectionSelectorIntent } from "types/gallery";
 import {
+    COLLECTION_SORT_ORDER,
+    CollectionSummaryType,
+    DUMMY_UNCATEGORIZED_COLLECTION,
     isAddToAllowedCollection,
     isMoveToAllowedCollection,
 } from "utils/collection";
