+++ web/apps/photos/src/components/Collections/CollectionShare/index.tsx
@@ -1,9 +1,9 @@
 import { EnteDrawer } from "@/base/components/EnteDrawer";
 import { Titlebar } from "@/base/components/Titlebar";
 import { DialogProps, Stack } from "@mui/material";
-import { CollectionSummaryType } from "constants/collection";
 import { t } from "i18next";
 import { Collection, CollectionSummary } from "types/collection";
+import { CollectionSummaryType } from "utils/collection";
 import EmailShare from "./emailShare";
 import PublicShare from "./publicShare";
 import SharingDetails from "./sharingDetails";
