+++ web/apps/photos/src/components/Collections/index.tsx
@@ -6,12 +6,13 @@ import CollectionListBar from "components/Collections/CollectionListBar";
 import { SetCollectionNamerAttributes } from "components/Collections/CollectionNamer";
 import CollectionShare from "components/Collections/CollectionShare";
 import { ITEM_TYPE, TimeStampListItem } from "components/PhotoList";
-import { ALL_SECTION, COLLECTION_LIST_SORT_BY } from "constants/collection";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { sortCollectionSummaries } from "services/collectionService";
 import { Collection, CollectionSummaries } from "types/collection";
 import { SetFilesDownloadProgressAttributesCreator } from "types/gallery";
 import {
+    ALL_SECTION,
+    COLLECTION_LIST_SORT_BY,
     hasNonSystemCollections,
     isSystemCollection,
     shouldBeShownOnCollectionBar,
