+++ web/apps/photos/src/components/PhotoFrame.tsx
@@ -8,7 +8,7 @@ import { CustomError } from "@ente/shared/error";
 import useMemoSingleThreaded from "@ente/shared/hooks/useMemoSingleThreaded";
 import { styled } from "@mui/material";
 import PhotoViewer from "components/PhotoViewer";
-import { TRASH_SECTION } from "constants/collection";
+import { TRASH_SECTION } from "utils/collection";
 import { useRouter } from "next/router";
 import { GalleryContext } from "pages/gallery";
 import PhotoSwipe from "photoswipe";
