+++ web/apps/photos/src/components/pages/gallery/PreviewCard.tsx
@@ -12,12 +12,12 @@ import {
     LoadingThumbnail,
     StaticThumbnail,
 } from "components/PlaceholderThumbnails";
-import { TRASH_SECTION } from "constants/collection";
 import { GAP_BTW_TILES, IMAGE_CONTAINER_MAX_WIDTH } from "constants/gallery";
 import i18n from "i18next";
 import { DeduplicateContext } from "pages/deduplicate";
 import { GalleryContext } from "pages/gallery";
 import React, { useContext, useEffect, useRef, useState } from "react";
+import { TRASH_SECTION } from "utils/collection";
 import { shouldShowAvatar } from "utils/file";
 import Avatar from "./Avatar";
