+++ web/apps/photos/src/components/pages/gallery/SelectedFileOptions.tsx
@@ -13,11 +13,6 @@ import UnArchiveIcon from "@mui/icons-material/Unarchive";
 import VisibilityOffOutlined from "@mui/icons-material/VisibilityOffOutlined";
 import VisibilityOutlined from "@mui/icons-material/VisibilityOutlined";
 import { Box, IconButton, Stack, Tooltip } from "@mui/material";
-import {
-    ALL_SECTION,
-    ARCHIVE_SECTION,
-    TRASH_SECTION,
-} from "constants/collection";
 import { t } from "i18next";
 import { AppContext } from "pages/_app";
 import { useContext } from "react";
@@ -26,7 +21,12 @@ import {
     CollectionSelectorIntent,
     SetCollectionSelectorAttributes,
 } from "types/gallery";
-import { COLLECTION_OPS_TYPE } from "utils/collection";
+import {
+    ALL_SECTION,
+    ARCHIVE_SECTION,
+    COLLECTION_OPS_TYPE,
+    TRASH_SECTION,
+} from "utils/collection";
 import { FILE_OPS_TYPE } from "utils/file";
 import { formatNumber } from "utils/number/format";
 import { getTrashFilesMessage } from "utils/ui";
